@@ -1,8 +1,15 @@
 import './app.css';
 
 import React from 'react';
-import ReactDOM from 'react-dom';
-
 import Form from './components/Form';
 
-ReactDOM.render(<Form />, document.getElementById('editor'));
+
+// For React < 18
+// import ReactDOM from 'react-dom';
+// ReactDOM.render(<Form />, document.getElementById('editor'));
+
+
+import { createRoot } from 'react-dom/client';
+const container = document.getElementById('editor');
+const root = createRoot(container);
+root.render(<Form/>);
